コマンドを分けてInterfaceのポリモーフィズム
