--- MinecraftDatapackReloadHelper/Interfaces/Commands/ToolCommand/IHaveArgs.cs
@@ -1,7 +1,0 @@
-﻿namespace MinecraftDatapackReloadHelper.Interfaces.Commands.ToolCommand
-{
-    internal interface IHaveArgs : IToolCommand
-    {
-        protected Task Run(List<string> args);
-    }
-}
