+++ MinecraftDatapackReloadHelper/Systems/Commands/Terminal.cs
@@ -1,0 +1,9 @@
+﻿using MinecraftDatapackReloadHelper.Interfaces.Commands;
+
+namespace MinecraftDatapackReloadHelper.Systems.Commands
+{
+    internal class Terminal : IToolCommand
+    {
+        public async Task Run(List<string> args) => await Control.Terminal.RunAsync();
+    }
+}
